Consume Nufi App Acta constitutiva

+++ Models/PersonaMoral.cs
@@ -4,7 +4,7 @@ namespace Nufi.kyb.v1.Models
 	{
         public string razon_social { get; set; }
         public string rfc { get; set; }
-        public int participacion_social { get; set; }
+        public double participacion_social { get; set; }
         public string giro_mercantil { get; set; }
         public string nacionalidad { get; set; }
         public PersonaFisica[] socios { get; set; }
